+++ lib/random_math_quiz.dart
@@ -2,6 +2,8 @@ library;
 
 export 'src/interface/math_quiz.dart';
 export 'src/domain/entities/math_question.dart';
+export 'src/domain/entities/expression.dart';
+export 'src/domain/repositories/math_question_repository.dart';
 export 'src/domain/value_objects/rational.dart';
 export 'src/domain/value_objects/math_op.dart';
 export 'src/domain/value_objects/difficulty.dart';